feat: complete figure area calculation function

+++ docker-entrypoint.sql
@@ -1,8 +1,8 @@
 CREATE TABLE results (
     id SERIAL PRIMARY KEY,
-    x INT,
-    y INT,
-    r INT,
+    x NUMERIC,
+    y NUMERIC,
+    r NUMERIC,
     ishit BOOLEAN,
     executiontime VARCHAR(32),
     currenttime TIMESTAMP
